tshirt_fulfillment/requirements.txt
# Requirements for T-shirt Fulfillment AI Agent

# Core dependencies
langchain>=0.0.267
redis>=4.5.1
ollama>=0.1.0
fastapi>=0.100.0
uvicorn>=0.22.0

# For local LLM integration
langchain-community>=0.0.10

# For Excel file handling
openpyxl>=3.1.0

# For Google Drive integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# For image generation (Stable Diffusion)
diffusers>=0.19.0
transformers>=4.30.0
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0